[refactor] RacingGameStatus 합치기
